chore: remove interface
